Create ProDocker.md
I have attached text files that use typescript and node.js to allow a command line driven login to AWS through an Azure portal.  Can you generate from these files a python equivalent, utilising the best features of python to improve the code when suitable.  Can the solution be presented with a Dockerfile that can be used to encapsulate all of the requirements and be run as if running a native binary command.  Feel free to use an alternative browser such as playright or selenium in your solution.